+++ test/index.html
@@ -3,17 +3,13 @@
 <html>
     <head>
         <title>Testing</title>
+        <link rel="stylesheet" type="text/css", href="./local/index.css" />
     </head>
     <body>
-        <div style="height: 200vh"> </div>
-        <div id="scrollable" style="background: black; overflow-y: scroll; height: 80vh">
-            <div style="height: 200vh"> </div>
-            <div style="background: blue; height: 50vh" id="element"> </div>
-            <div style="height: 400vh"> </div>
+        <div id="scrollable">
+            <div id="element"> </div>
         </div>
-        <div style="height: 400vh"> </div>
-        <div style="background: yellow; height: 80vh" id="element1"> </div>
-        <div style="height: 400vh"> </div>
-        <script src="./loadScroll.ts"> </script>
+        <div id="element1"> </div>
     </body>
+    <script src="./loadScroll.ts"> </script>
 </html>
